requirements.txt
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/altgraph-0.17.2-py2.py3-none-any.whl
asgiref==3.6.0
beautifulsoup4==4.12.2
boto3==1.28.49
botocore==1.31.49
crispy-bootstrap5==0.7
Django==4.1.7
django-bootstrap-v5==1.0.11
django-components==0.28.2
django-crispy-forms==2.0
django-storages==1.14
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/future-0.18.2-py3-none-any.whl
jmespath==1.0.1
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/macholib-1.15.2-py2.py3-none-any.whl
Pillow==10.0.1
psycopg2-binary==2.9.7
python-dateutil==2.8.2
s3transfer==0.6.2
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/six-1.15.0-py2.py3-none-any.whl
soupsieve==2.5
sqlparse==0.4.3
urllib3==1.26.16
